final/web/Baby Flask/service/requirements.txt
flask==2.3.2
pyjwt==2.7.0
